refactor(checker): split oversized call checker helpers
