+++ crates/tsz-checker/src/checkers/call_checker/candidate_collection.rs
@@ -29,7 +29,7 @@ const SPREAD_ARGUMENT_MARKER_NAME: &str = "__tsz_spread_argument__";
 /// tuple. A *fixed*-length tuple constraint (`[number, number]`) stays
 /// positionally expandable, so multi-spread combos (`f(...u, ...v)`) keep
 /// inferring as before.
-fn type_param_variadic_tuple_spread(
+pub(super) fn type_param_variadic_tuple_spread(
     db: &dyn tsz_solver::construction::TypeDatabase,
     spread_type: TypeId,
     elems: &[TupleElement],
@@ -1849,164 +1849,6 @@ impl<'a> CheckerState<'a> {
                 self.check_object_literal_excess_properties(arg_type, expected, arg_idx);
             }
         }
-    }
-
-    pub(super) fn validate_non_tuple_spreads_for_signature(
-        &mut self,
-        args: &[NodeIndex],
-        func_type: TypeId,
-    ) {
-        if let Some(arg_idx) = self.first_non_tuple_spread_rejected_by_signature(args, func_type) {
-            self.error_spread_must_be_tuple_or_rest_at(arg_idx);
-        }
-    }
-
-    /// Like [`Self::validate_non_tuple_spreads_for_signature`] but returns the
-    /// first spread argument that `func_type` cannot absorb at a rest (or
-    /// optional-tail) position, without emitting a diagnostic. Overload
-    /// resolution uses this to treat a non-tuple spread that overflows a
-    /// fixed-arity overload as a *soft* mismatch (try the next overload) rather
-    /// than committing TS2556 — a sibling overload with a rest parameter may
-    /// absorb the spread (#14319).
-    pub(super) fn first_non_tuple_spread_rejected_by_signature(
-        &mut self,
-        args: &[NodeIndex],
-        func_type: TypeId,
-    ) -> Option<NodeIndex> {
-        let ctx = ContextualTypeContext::with_expected(self.ctx.types, func_type);
-        let mut effective_index = 0usize;
-        for &arg_idx in args {
-            let Some(arg_node) = self.ctx.arena.get(arg_idx) else {
-                effective_index += 1;
-                continue;
-            };
-            if arg_node.kind != syntax_kind_ext::SPREAD_ELEMENT {
-                effective_index += 1;
-                continue;
-            }
-            let Some(spread_data) = self.ctx.arena.get_spread(arg_node) else {
-                effective_index += 1;
-                continue;
-            };
-            let spread_type = self.normalized_spread_argument_type(spread_data.expression);
-            // A variadic-tuple type-parameter spread stays a single unit (see argument
-            // collection and the type-parameter spread branch below); do not
-            // advance by its constraint's tuple element count.
-            if let Some(elems) = tuple_elements_for_type(self.ctx.types, spread_type)
-                && !type_param_variadic_tuple_spread(self.ctx.types, spread_type, &elems)
-            {
-                effective_index += elems.len();
-                continue;
-            }
-            // An array literal spread (e.g. `...['a', 'x']`) is expanded element-by-element
-            // during argument collection, so each element is checked individually against
-            // the corresponding parameter. Treat it like a tuple-like spread here: advance
-            // by the literal's element count and skip the TS2556 emission. tsc behaves the
-            // same way — TS2556 is only reported for spreads of opaque arrays/iterables
-            // whose runtime length is unknown at the call site.
-            if array_element_type_for_type(self.ctx.types, spread_type).is_some()
-                && let Some(expr_node) = self
-                    .ctx
-                    .arena
-                    .get(self.ctx.arena.skip_parenthesized(spread_data.expression))
-                && let Some(literal) = self.ctx.arena.get_literal_expr(expr_node)
-            {
-                effective_index += literal.elements.nodes.len();
-                continue;
-            }
-            if is_type_parameter_type(self.ctx.types, spread_type)
-                && let Some(constraint) = crate::query_boundaries::common::type_parameter_constraint(
-                    self.ctx.types,
-                    spread_type,
-                )
-                && (array_element_type_for_type(self.ctx.types, constraint).is_some()
-                    || tuple_elements_for_type(self.ctx.types, constraint).is_some())
-            {
-                effective_index += 1;
-                continue;
-            }
-            let is_non_tuple_spread = array_element_type_for_type(self.ctx.types, spread_type)
-                .is_some()
-                || self.is_iterable_type(spread_type);
-            if is_non_tuple_spread && !ctx.allows_non_tuple_spread_position(effective_index) {
-                return Some(arg_idx);
-            }
-            effective_index += 1;
-        }
-        None
-    }
-
-    pub(super) fn find_prior_non_tuple_spread_for_mismatch(
-        &mut self,
-        args: &[NodeIndex],
-        mismatch_index: usize,
-    ) -> Option<NodeIndex> {
-        let mut effective_index = 0usize;
-        let mut prior_non_tuple_spread = None;
-
-        for &arg_idx in args {
-            if effective_index > mismatch_index {
-                break;
-            }
-            let Some(arg_node) = self.ctx.arena.get(arg_idx) else {
-                effective_index += 1;
-                continue;
-            };
-            if arg_node.kind != syntax_kind_ext::SPREAD_ELEMENT {
-                if effective_index == mismatch_index {
-                    return prior_non_tuple_spread;
-                }
-                effective_index += 1;
-                continue;
-            }
-            let Some(spread_data) = self.ctx.arena.get_spread(arg_node) else {
-                effective_index += 1;
-                continue;
-            };
-            let spread_type = self.normalized_spread_argument_type(spread_data.expression);
-            // A variadic-tuple type-parameter spread stays a single unit (see argument
-            // collection); treat it as one position rather than its constraint's
-            // tuple element count.
-            if let Some(elems) = tuple_elements_for_type(self.ctx.types, spread_type)
-                && !type_param_variadic_tuple_spread(self.ctx.types, spread_type, &elems)
-            {
-                if mismatch_index < effective_index + elems.len() {
-                    return prior_non_tuple_spread;
-                }
-                effective_index += elems.len();
-                continue;
-            }
-            // An array literal spread (e.g. `...['a', 'x']`) is expanded element-by-element
-            // during argument collection. A mismatch at one of those expanded indices is a
-            // per-element type error (TS2345/TS2322), not a TS2556. Skip past the literal's
-            // elements without setting `prior_non_tuple_spread`.
-            if array_element_type_for_type(self.ctx.types, spread_type).is_some()
-                && let Some(expr_node) = self
-                    .ctx
-                    .arena
-                    .get(self.ctx.arena.skip_parenthesized(spread_data.expression))
-                && let Some(literal) = self.ctx.arena.get_literal_expr(expr_node)
-            {
-                let count = literal.elements.nodes.len();
-                if mismatch_index < effective_index + count {
-                    return prior_non_tuple_spread;
-                }
-                effective_index += count;
-                continue;
-            }
-            let is_non_tuple_spread = array_element_type_for_type(self.ctx.types, spread_type)
-                .is_some()
-                || self.is_iterable_type(spread_type);
-            if effective_index == mismatch_index {
-                return prior_non_tuple_spread;
-            }
-            if is_non_tuple_spread {
-                prior_non_tuple_spread = Some(arg_idx);
-            }
-            effective_index += 1;
-        }
-
-        prior_non_tuple_spread
     }
 }
