+++ crates/tsz-checker/src/checkers/call_checker/overload_resolution.rs
@@ -4,6 +4,7 @@
 
 mod contextual_retry;
 mod helpers;
+mod mismatch_helpers;
 mod return_context;
 
 use crate::context::TypingRequest;
@@ -1988,38 +1989,6 @@ impl<'a> CheckerState<'a> {
                 fallback_return,
             },
             selected_type_predicate: None,
-        })
-    }
-
-    fn recheck_overload_args_after_mismatch_without_context(
-        &mut self,
-        args: &[NodeIndex],
-        mismatch_index: usize,
-    ) {
-        for &arg_idx in args.iter().skip(mismatch_index.saturating_add(1)) {
-            if !self.is_callback_like_argument(arg_idx) {
-                continue;
-            }
-
-            for callback_idx in self.callback_function_indices(arg_idx) {
-                self.ctx
-                    .implicit_any_contextual_closures
-                    .remove(&callback_idx);
-                self.ctx.implicit_any_checked_closures.remove(&callback_idx);
-            }
-            self.invalidate_expression_for_contextual_retry(arg_idx);
-            let _ = self.get_type_of_node_with_request(arg_idx, &TypingRequest::NONE);
-        }
-    }
-
-    fn arg_source_span(&self, args: &[NodeIndex], index: usize) -> Option<tsz_solver::SourceSpan> {
-        let &arg_idx = args.get(index)?;
-        self.ctx.arena.get(arg_idx).map(|node| {
-            tsz_solver::SourceSpan::new(
-                self.ctx.file_name.as_str(),
-                node.pos,
-                node.end.saturating_sub(node.pos),
-            )
         })
     }
 }
